manager.py contains manager() func to scrape api/websites and control arduino (not created yet).
The function manager() runs first webscrapes api.weather.gov using json.
Then it process the data and stores it in a csv. Also, it runs the
webscraper (scrapy) and it webscraped weather.gov, then it process the
data and stores it in a csv.

The function manager() does not yet run arduino code.